Refactoring

+++ src/main/java/ru/job4j/chat/service/MessageService.java
@@ -5,6 +5,7 @@ import ru.job4j.chat.model.Message;
 import ru.job4j.chat.repository.MessageRepository;
 
 import java.util.Date;
+import java.util.NoSuchElementException;
 
 @Service
 public class MessageService {
@@ -21,6 +22,9 @@ public class MessageService {
     }
 
     public void deleteById(int id) {
+        if (repository.findById(id).isEmpty()) {
+            throw new NoSuchElementException("Message not found");
+        }
         repository.deleteById(id);
     }
 }
